requirements.txt
streamlit
pandas
numpy
scikit-learn
opencv-python
mahotas
matplotlib
seaborn
Pillow
scikit-image
joblib
